Refactord factories.py into factories folder with files containing the factories
