+++ api/app/tests/apps/signals/factories/type.py
@@ -1,0 +1,11 @@
+import factory
+
+from signals.apps.signals.models import Type
+
+
+class TypeFactory(factory.DjangoModelFactory):
+    _signal = factory.SubFactory('tests.apps.signals.factories.signal.SignalFactory')
+    name = Type.SIGNAL  # Default type is a "Signal" (Melding in Dutch)
+
+    class Meta:
+        model = Type
